moved ODM options WTForms to separate file #34
